added a new APIBlockProcessor (of type API_EXTENSION_BLOCK) that allows users to include extension blocks with bundles generated at the API
added a BlockInfo list to the Bundle class to store the extension blocks
generated at the API

+++ servlib/bundling/BlockInfo.h
@@ -77,7 +77,7 @@ public:
 
 protected:
     BlockProcessor*  owner_;       ///< Owner of this block
-    u_int8_t	     owner_type_;  ///< Extracted from owner
+    u_int16_t	     owner_type_;  ///< Extracted from owner
     const BlockInfo* source_;      ///< Owner of this block
     DataBuffer       contents_;    ///< Block contents with length set to
                                    ///  the amount currently in the buffer
